Move all specs
